requirements/requirements-logfire.txt
# This file was autogenerated by uv via the following command:
#    uv pip compile requirements/requirements-logfire.in -o requirements/requirements-logfire.txt
annotated-types==0.7.0
    # via pydantic
anyio==4.11.0
    # via
    #   claude-agent-sdk
    #   httpx
    #   mcp
    #   sse-starlette
    #   starlette
attrs==25.4.0
    # via
    #   jsonschema
    #   referencing
certifi==2025.10.5
    # via
    #   httpcore
    #   httpx
    #   requests
charset-normalizer==3.4.4
    # via requests
claude-agent-sdk==0.1.3
    # via -r requirements/requirements.in
click==8.3.0
    # via
    #   typer
    #   uvicorn
executing==2.2.1
    # via logfire
googleapis-common-protos==1.70.0
    # via opentelemetry-exporter-otlp-proto-http
h11==0.16.0
    # via
    #   httpcore
    #   uvicorn
httpcore==1.0.9
    # via httpx
httpx==0.28.1
    # via mcp
httpx-sse==0.4.3
    # via mcp
idna==3.11
    # via
    #   anyio
    #   httpx
    #   requests
importlib-metadata==8.7.0
    # via opentelemetry-api
jsonschema==4.25.1
    # via mcp
jsonschema-specifications==2025.9.1
    # via jsonschema
logfire==4.13.2
    # via -r requirements/requirements-logfire.in
loguru==0.7.3
    # via -r requirements/requirements.in
markdown-it-py==4.0.0
    # via rich
mcp==1.17.0
    # via claude-agent-sdk
mdurl==0.1.2
    # via markdown-it-py
opentelemetry-api==1.37.0
    # via
    #   -r requirements/requirements.in
    #   opentelemetry-exporter-otlp-proto-http
    #   opentelemetry-instrumentation
    #   opentelemetry-sdk
    #   opentelemetry-semantic-conventions
opentelemetry-exporter-otlp-proto-common==1.37.0
    # via opentelemetry-exporter-otlp-proto-http
opentelemetry-exporter-otlp-proto-http==1.37.0
    # via
    #   -r requirements/requirements.in
    #   logfire
opentelemetry-instrumentation==0.58b0
    # via logfire
opentelemetry-proto==1.37.0
    # via
    #   opentelemetry-exporter-otlp-proto-common
    #   opentelemetry-exporter-otlp-proto-http
opentelemetry-sdk==1.37.0
    # via
    #   -r requirements/requirements.in
    #   logfire
    #   opentelemetry-exporter-otlp-proto-http
opentelemetry-semantic-conventions==0.58b0
    # via
    #   opentelemetry-instrumentation
    #   opentelemetry-sdk
packaging==25.0
    # via opentelemetry-instrumentation
protobuf==6.32.1
    # via
    #   googleapis-common-protos
    #   logfire
    #   opentelemetry-proto
pydantic==2.12.2
    # via
    #   -r requirements/requirements.in
    #   mcp
    #   pydantic-settings
pydantic-core==2.41.4
    # via pydantic
pydantic-settings==2.11.0
    # via mcp
pygments==2.19.2
    # via rich
python-dotenv==1.1.1
    # via
    #   -r requirements/requirements.in
    #   pydantic-settings
python-multipart==0.0.20
    # via mcp
referencing==0.37.0
    # via
    #   jsonschema
    #   jsonschema-specifications
requests==2.32.5
    # via opentelemetry-exporter-otlp-proto-http
rich==14.2.0
    # via
    #   -r requirements/requirements.in
    #   logfire
    #   typer
rpds-py==0.27.1
    # via
    #   jsonschema
    #   referencing
shellingham==1.5.4
    # via typer
sniffio==1.3.1
    # via anyio
sse-starlette==3.0.2
    # via mcp
starlette==0.48.0
    # via mcp
typer==0.19.2
    # via -r requirements/requirements.in
typing-extensions==4.15.0
    # via
    #   logfire
    #   opentelemetry-api
    #   opentelemetry-exporter-otlp-proto-http
    #   opentelemetry-sdk
    #   opentelemetry-semantic-conventions
    #   pydantic
    #   pydantic-core
    #   typer
    #   typing-inspection
typing-inspection==0.4.2
    # via
    #   pydantic
    #   pydantic-settings
urllib3==2.5.0
    # via requests
uvicorn==0.37.0
    # via mcp
wrapt==1.17.3
    # via opentelemetry-instrumentation
zipp==3.23.0
    # via importlib-metadata
